+++ lib/use_case/get_document_version_use_case.dart
@@ -5,11 +5,10 @@ import 'package:html/parser.dart' as html show parse;
 import 'package:http/http.dart' as http;
 import 'package:injectable/injectable.dart';
 
-/// Gets the document - Privacy Policy or Terms of Service - version.
-///
-/// Impl. reads value from "version" `meta` tag.
+/// Gets the HTML document version by reading value
+/// from "version" HTML `meta` tag.
 @lazySingleton
-class GetDocumentVersionUseCase {
+class GetHtmlDocumentVersionUseCase {
   static final http.Client _client = http.Client();
 
   /// Gets the document version on this [url].
@@ -21,17 +20,7 @@ class GetDocumentVersionUseCase {
           url, "url", "GET on URL yields no text content.");
     }
 
-    final document = html.parse(text);
-    final meta =
-        document.querySelector('html > head > meta[name="version"][content]');
-    final version = (meta?.attributes["content"]?.trim() ?? '');
-
-    if (version.isEmpty) {
-      throw ArgumentError.value(
-          url, "url", "Required meta tag is not present.");
-    }
-
-    return version;
+    return parseVersion(text);
   }
 
   /// Gets the HTML content from given [url].